Java 7 complains about this type

+++ dd-smoke-tests/cli/src/test/groovy/datadog/smoketest/CliApplicationSmokeTest.groovy
@@ -5,7 +5,7 @@ import java.util.concurrent.TimeUnit
 
 class CliApplicationSmokeTest extends AbstractSmokeTest {
   // Estimate for the amount of time instrumentation, plus request, plus some extra
-  private static final int TIMEOUT_SECS = 10
+  private static final long TIMEOUT_SECS = 10
 
   @Override
   ProcessBuilder createProcessBuilder() {
